# openapi.yml
openapi: 3.1.2
info:
  title: "短网址"
  description: "短网址接口"
  contact:
    name: "Jetsung Chan"
    url: "https://github.com/jetsung"
    email: "i@jetsung.com"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.3.0-preview.1
servers:
  - url: http://127.0.0.1:8080
    description: Version 1
...
paths:
...
  /api/histories:
    get:
      tags:
        - history
      summary: "获取访问历史记录"
      description: "获取所有访问日志信息"
      operationId: "getHistories"
      parameters:
        - name: page
          in: query
          description: "页码"
          required: false
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: per_page
          in: query
          description: "每页条数"
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: sort_by
          in: query
          description: "排序字段"
          required: false
          schema:
            type: string
            default: "accessed_at"
            enum:
              - id
              - accessed_at
              - created_at
        - name: order
          in: query
          description: "排序方向"
          required: false
          schema:
            type: string
            default: "desc"
            enum:
              - asc
              - desc
        - name: short_code
          in: query
          description: "短码搜索"
          required: false
          schema:
            type: string
            maxLength: 16
        - name: ip_address
          in: query
          description: "IP地址搜索"
          required: false
          schema:
            type: string
      responses:
        "200":
          description: "操作成功"
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/HistoryResponse"
                  meta:
                    $ref: "#/components/schemas/PageMeta"
        "400":
          description: "请求错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/histories/batch-delete:
    post:
      tags:
        - history
      summary: "批量删除访问日志"
      description: "批量删除访问日志"
      operationId: "deleteHistories"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchDeleteRequest"
      responses:
        "204":
          description: "操作成功"
        "400":
          description: "请求错误"
        "404":
          description: "日志不存在"
        "500":
          description: "操作失败"
        default:
          description: "未知错误"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
    PageMeta:
      type: object
      required:
        - page
        - per_page
        - count
        - total
        - total_pages
      properties:
        page:
          type: integer
          minimum: 1
          description: 当前页码
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          description: 每页数量
          example: 10
        count:
          type: integer
          minimum: 0
          description: 当前页条目数
          example: 10
        total:
          type: integer
          minimum: 0
          description: 总条目数
          example: 100
        total_pages:
          type: integer
          minimum: 0
          description: 总页数
          example: 10
...
    ErrorResponse:
      required:
        - error_code
      type: object
      properties:
        error_code:
          type: string
          description: 业务约定的错误码
        error_message:
          type: string
          description: 错误信息
...
    BatchDeleteRequest:
      type: object
      required:
        - ids
      properties:
        ids:
          type: array
          items:
            type: integer
            format: int64
          description: "要删除的 ID 列表"
          example: [1, 2, 3, 4, 5]
...
    HistoryResponse:
      type: object
      required:
        - id
        - url_id
        - short_code
        - ip_address
        - user_agent
        - accessed_at
        - created_at
      properties:
        id:
          type: integer
          format: int64
          minimum: 1
          description: 访问记录 ID
        url_id:
          type: integer
          format: int32
          description: 短链接 ID
        short_code:
          type: string
          pattern: ^[a-zA-Z0-9]{5,16}$
          example: "xY7z9"
          description: 短码
        ip_address:
          type: string
          format: ipv4
          example: "192.168.0.1"
          description: 访问者 IP
        user_agent:
          type: string
          minLength: 10
          example: "Mozilla/5.0 (Windows NT 10.0)"
          description: 浏览器 UA
        referer:
          anyOf:
            - type: string
              format: uri
            - type: "null"
          example: "https://google.com"
          description: 来源页面
        country:
          type: string
          example: "中国"
        region:
          type: string
          example: "华南"
        province:
          type: string
          example: "广东省"
        city:
          type: string
          example: "深圳市"
        isp:
          type: string
          example: "中国电信"
        device_type:
          type: string
          enum: [pc, mobile, tablet]
        os:
          type: string
          enum: [Windows, MacOS, Linux, Android, iOS]
        browser:
          type: string
          enum: [Chrome, Firefox, Safari, Edge, IE]
        accessed_at:
          type: string
          format: date-time
          description: "访问时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"
        created_at:
          type: string
          format: date-time
          description: "记录创建时间 (ISO 8601 格式)"
          example: "2024-01-15T08:30:00Z"